First brief edit of first three sections of Fuel docs
Editing docs for spelling and some grammatical errors.  Stuck to the
more obvious things for now.  The first three sections are ready: Basic,
Installation, General. Also, renamed Orm to ORM to stick to proper
acronym rules. May not fly but whatevs.  Tis your prerogative.   Also,
going forward, all headings should have proper capitalization.  All
words in headings should be capitalized except for coordination
conjunctions. So, words like a, an, and, but, or, nor, so and so on. You
guys might want to stick to that in the future, but again, tis your
prerogative.   Hope this is cool with you guys.   Thanks.