Section: General
When it came to capitalizing model, view or controller in context as a
Fuel entity, my approach was this:   If I can take that sentence and
move it into any MVC framework's documentation and it still made sense
then it was not specific enough, and therefore, I did not capitalize it.
It was difficult and confusing at times so that was the approach that
seemed most logical to me.