Fixed some more typos, added a little clarity (see further description)
In classes/file/usage.html the File::download method description just
had 'download a file'. To me that would mean that it fetches a remote
URL to a local file, but actually it's a function that offers a local
file as a download to the website user. Thus I amended the description
here, it's not perfect but I think it's better than the previous
description.