Update .gitignore
There are some settings where you can have Vim create a swap file. Basically, it's a temporary version of the file that tracks the changes made within a single file so that you can actually undo past the last opening of the file. It's a good idea for Git to ignore.